Fix imports and cypress tests
