+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/api/outputs/AlertDataOutput.kt
@@ -8,6 +8,10 @@ class AlertDataOutput(
     val seaFront: String? = null,
     val dml: String? = null,
     val natinfCode: Int? = null,
+    val riskFactor: Double? = null,
+    val alertId: Int? = null,
+    val name: String,
+    val description: String? = null,
 ) : ReportingValueDataOutput() {
     companion object {
         fun fromAlertType(alert: Alert) =
@@ -16,6 +20,10 @@ class AlertDataOutput(
                 seaFront = alert.seaFront,
                 dml = alert.dml,
                 natinfCode = alert.natinfCode,
+                riskFactor = alert.riskFactor,
+                alertId = alert.alertId,
+                name = alert.name,
+                description = alert.description,
             )
     }
 }
