+++ frontend/src/features/Alert/constants.ts
@@ -157,3 +157,26 @@ export const ALERTS_ARCHIVED_AFTER_NEW_VOYAGE: string[] = Object.keys(COMMON_ALE
   .map(alertTypeName => COMMON_ALERT_TYPE_OPTION[alertTypeName])
   .filter(alertType => alertType.isArchivedAfterDEPMessage)
   .map(alert => alert.code)
+
+export enum AdministrativeAreaType {
+  DISTANCE_TO_SHORE = 'DISTANCE_TO_SHORE',
+  EEZ_AREA = 'EEZ_AREA',
+  FAO_AREA = 'FAO_AREA',
+  NEAFC_AREA = 'NEAFC_AREA'
+}
+
+export const AdministrativeAreaTypeLabel: Record<AdministrativeAreaType, string> = {
+  DISTANCE_TO_SHORE: 'Distances à la côte',
+  EEZ_AREA: 'Zones ZEE',
+  FAO_AREA: 'Zones FAO',
+  NEAFC_AREA: 'Zones NEAFC'
+}
+
+export enum AdministrativeAreaValueLabel {
+  '0-12_MINUS_DE_FISHING_AREA' = '12 milles (sans la ZEE DE)',
+  '0-12_MINUS_ES_FISHING_AREA' = '12 milles (sans la ZEE ESP)',
+  '0-3' = '3 milles',
+  '12_MINUS_BE_AND_NL_FISHING_AREAS' = '12 milles (sans les ZEEs BEL et NLD)',
+  '3-6' = 'Entre 3 et 6 milles',
+  '6-12' = 'Entre 6 et 12 milles'
+}
